# Speedup Pretty Print